Finished wiring up Spring data and JPA

+++ src/main/java/fi/pizzalovers/blogbackend/MyRepo.java
@@ -1,6 +1,5 @@
 package fi.pizzalovers.blogbackend;
 
 import org.springframework.data.repository.CrudRepository;
-
-public interface MyRepo extends CrudRepository<BlogPost, Integer> {
+public interface MyRepo extends CrudRepository<BlogPost, Long> {
 }
